\name{print.Gls.ret}
\alias{print.Gls.ret}

\title{Printing the detection result}

\description{ 
 Print the result information.
} 

\usage{ print (r.sum.ret)}

\arguments{
\item{r.sum.ret}{ List, the class is sum.BLS.ret.}
}

\details{

The print command will print out the result object in the standard format. 
}

\author{Zhong Wang, \email{zzw2@psu.edu} }

\examples{
r.gls;
}

\keyword{print}