\name{print.sum.BLS.ret}
\alias{print.sum.BLS.ret}

\title{Printing the detection result}

\description{ 
 Print the result information.
} 

\usage{ 
\method{print}{sum.BLS.ret}(x, \dots) 
}

\arguments{
\item{x}{An object return by \code{\link{bls.simple}} or \code{\link{bls.plink}}or \code{\link{bls.snpmat}}}
\item{...}{Other parameters}
}

\details{

The print command will print out the result object in the standard format. 
}

\author{Zhong Wang, \email{zzw2@psu.edu} }

\seealso{See \code{\link{bls.simple}} or \code{\link{bls.plink}}or \code{\link{bls.snpmat}} to generate an object }

\keyword{print}