\name{print.sum.GLS.ret}
\alias{print.sum.GLS.ret}


\title{Printing the detection result}

\description{ 
 Print the result information.
} 

\usage{ 
\method{print}{sum.GLS.ret}(x, \dots) 
}

\arguments{
\item{x}{An object return by \code{\link{gls.simple}} or \code{\link{gls.plink}} or \code{\link{gls.snpmat}}}
\item{...}{Other parameters}
}

\details{

The print command will print out the result object in the standard format. 
}

\author{Zhong Wang, \email{zzw2@psu.edu} }

\seealso{See \code{\link{gls.simple}} or \code{\link{gls.plink}}or \code{\link{gls.snpmat}} to generate an object }


\keyword{print}