\name{bls.simulate}
\alias{bls.simulate}

\title{ Simulation data of the BLS model}

\description{ 
Generating simulation data object by the BLS model.
} 

\usage{
bls.simulate(phe.out,
      snp.out,
      simu_grp=1,
      simu_n= 500,
      simu_p=1000, 
      simu_snp_rho = 0.1, 
      simu_snp_missing = 0.002, 
      simu_rho = 0.4, 
      simu_sigma2 = 3, 
      simu_mu = 26, 
      simu_cov_range = c(0,1),
      simu_cov_effect= c( 0, 2 ), 
      simu_add_pos   = c( 100, 200, 300), 
      simu_add_effect= c(2.2,-2.5,2.0), 
      simu_dom_pos = c(300, 500, 700), 
      simu_dom_effect= c(2.8,2.0,-2.5), 
      simu_t_range = c(-1,1), debug=F)
}

\arguments{
\item{phe.out}{ String, the name of output phenotypic data file. }
\item{snp.out}{ String, the name of output genotypic data file. }
\item{simu_grp}{ Numeric, the number of groups that snps divided into in the parallel computation. Default: 1.}
\item{simu_n}{ Integer, sample size. Default: 500. }
\item{simu_p}{ Integer, number of SNPs. Default: 1000. }
\item{simu_snp_rho}{ Float, correlation coefficient between two adjacent SNPs. Default: 0.1.}
\item{simu_snp_missing}{Float, ratio of missing SNPs.}
\item{simu_rho}{ Float, correlation coefficient between two adjacent time points. Default: 0.4.}
\item{simu_sigma2}{ Float, residual error. Default: 3.}
\item{simu_mu}{ Float, overall mean of phenotypic data. Default: 26. }
\item{simu_cov_range}{ Vector, range of covariates. Default: (0,1).}
\item{simu_cov_effect}{ A vector indicating coefficient between two covariates. Default: (0,2).}
\item{simu_add_pos}{ Vector, positions of the significant SNPs with additive effects. \cr Default: (100, 200, 300).}
\item{simu_add_effect}{ Vector, additive effects of significant SNPs. Default: (2.2, -2.5, 2).}
\item{simu_dom_pos}{ Vector, positions of the significant SNPs with dominant effects. \cr Default: (300, 500,  700).}
\item{simu_dom_effect}{ Vector, dominant effects of significant SNPs. Default: (2.8, 2, -2.5). }
\item{simu_t_range}{ Vector, range of time points. Default: (-1, 1) }
\item{debug}{ Logical, if set to TRUE, printing error messages. Default: FALSE. }
}

\value{
Returns two simulation data files: the phenotypic data file and the genotypic data file.
}

\details{
The simulation in this package uses the pre-defined parameters to create a data object
containing phenotypic data wiht one single time point and genotypic data. 
The genotypic data file is coded 2, 1, 0, -1 to indicate genotypes QQ, Qq, qq and missing data respectively.\cr

It is assumed that the trait is controlled by some SNPs, and the positions of these SNPs will be generated 
by the parameters \strong{simu_a_pos} and \strong{simu_d_pos}. The genetic effects of overall mean and causal 
SNPs will be generated by the parameters \strong{simu_mu}, \strong{simu_a_effect} and \strong{simu_d_effect}.\cr

Given phenotypic data and genotype information, genetic effects of each SNPs could be estimated. However in GWAS 
a number of covariates either discrete or continuous may be measured for each subject. In the simulation function, 
covariate effects are generated by parameter \strong{simu_covar_effect}.\cr
}

\author{Zhong Wang and Nating Wang }

\references{
 Li, J., Das, K., Fu, G., Li, R., & Wu, R. (2011). The Bayesian lasso for genome-wide association studies. Bioinformatics, 27(4), 516-523.
}

\examples{
bls.simulate(phe.out="bls.simple.phe", snp.out="bls.simple.snp")
}

\keyword{functions}
\keyword{BLS model}
\keyword{simulate}


